# tests/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "MeDetail": {
        "additionalProperties": false,
        "description": "The current user's own record.\n\nImplements: DTO.\n\nAttributes:\n    id: The user's id.\n    display_name: The optional display name, shown only to the user themself.\n    roles: Platform roles held explicitly.\n    memberships: Memberships of active organisations, ordered by slug.\n    version: Optimistic-concurrency version, for ``ETag``.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/MembershipSummary"
            },
            "maxItems": 1000,
            "title": "Memberships",
            "type": "array"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "maxItems": 6,
            "title": "Roles",
            "type": "array",
            "uniqueItems": true
          },
          "version": {
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name",
          "roles",
          "memberships",
          "version"
        ],
        "title": "MeDetail",
        "type": "object"
      },
...
      "UpdateMeRequest": {
        "additionalProperties": false,
        "description": "Body of ``PATCH /api/v1/me``.\n\nImplements: API Schema.\n\nAttributes:\n    display_name: The new display name, 1 to 120 characters after NFC and\n        stripping, or ``null`` to clear it. Required, so clearing is explicit.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "UpdateMeRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "An OpenID Connect access token from the Yakhnama identity provider.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Yakhnama",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "description": "Return the caller's own record, mirroring them on first sight.\n\nArgs:\n    actor: The authenticated caller.\n    response: Used to set the ``ETag`` header.\n    services: Identity services bound by the composition root.\n\nReturns:\n    The caller's record with roles and memberships.",
        "operationId": "read_me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read Me",
        "tags": [
          "identity"
        ]
      },
      "patch": {
        "description": "Set or clear the caller's display name if they hold the current version.\n\nArgs:\n    body: The new display name.\n    actor: The authenticated caller.\n    response: Used to set the new ``ETag``.\n    services: Identity services bound by the composition root.\n    if_match: The ETag of ``GET /me`` the change is based on.\n\nReturns:\n    The caller's updated record.\n\nRaises:\n    PreconditionRequiredError: Without ``If-Match``.\n    PreconditionFailedError: If ``If-Match`` is stale or names another user.",
        "operationId": "update_me_api_v1_me_patch",
        "parameters": [
          {
            "description": "The resource's ETag, for example \"<id>:3\".",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The resource's ETag, for example \"<id>:3\".",
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "412": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "428": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/api/v1/organizations": {
      "post": {
        "description": "Create an organisation whose creator becomes its first admin.\n\nArgs:\n    body: Slug, name and type.\n    actor: The authenticated caller.\n    response: Used to set ``Location`` and ``ETag``.\n    services: Identity services bound by the composition root.\n    idempotency_key: Documented here; the idempotency middleware acts on it.\n\nReturns:\n    The created organisation.",
        "operationId": "create_organization_api_v1_organizations_post",
        "parameters": [
          {
            "description": "A UUID chosen by the client; replaying it with the same request returns the stored response.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID chosen by the client; replaying it with the same request returns the stored response.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Organization",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}